make Mosfet and VoltageSource inherit from new Device parent class
