+++ src/types.ts
@@ -24,13 +24,6 @@ export enum RelativeDirection {
 
 export type canvasId = 'main' | 'mosfet' | 'chart'
 
-export type BoundingBox = {
-  topLeft: TectonicPoint
-  topRight: TectonicPoint
-  bottomRight: TectonicPoint
-  bottomLeft: TectonicPoint
-}
-
 export type SchematicEffect = {
   node: Ref<Node>,
   origin: TectonicPoint,
